Extract handleMessage.js from server.js
